Move difficulty to model and auto launch config activity
