+++ app/src/main/java/neighbors/com/spacetrader/MainActivity.java
@@ -1,8 +1,10 @@
 package neighbors.com.spacetrader;
 
+import android.content.Intent;
 import android.os.Bundle;
 
 import androidx.appcompat.app.AppCompatActivity;
+import neighbors.com.spacetrader.ui.configuration.ConfigurationActivity;
 
 public class MainActivity extends AppCompatActivity {
 
@@ -10,5 +12,6 @@ public class MainActivity extends AppCompatActivity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
+        startActivity(new Intent(this, ConfigurationActivity.class));
     }
 }
